// src/types.ts
export interface ModelItem {
  id: string;
  title: string;
  image: string;
}

export interface BenefitItem {
  number: string;
  text: string;
}

export interface LampItem {
  id: number;
  image: string;
  title: string;
}

export interface GuideItem {
  id: string;
  title: string;
  originalPrice: string;
  image: string;
}

export interface FaqItem {
  question: string;
  answer: string;
}